refactor : RacingGame 리팩터링
RacingGame의 List<Cars>를 Cars 일급 컬렉션으로 대체
RacingGame의 Winners 일급 컬렉션을 Cars의 멤버 변수로 추가
